src/voice_agent/requirements.txt
# Voice Agent Dependencies
twilio>=8.0.0
elevenlabs>=1.0.0
openai-whisper>=20231117
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
